testing multiple objects
