+++ src/raytracing/render_utils.c
@@ -1,0 +1,40 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   render_uitls.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ssoeno <ssoeno@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/04/27 16:21:04 by ssoeno            #+#    #+#             */
+/*   Updated: 2025/04/27 16:27:34 by ssoeno           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "scene.h"
+#include <mlx_util.h>
+#include <raytracing.h>
+
+static void	ft_pixel_put(int x, int y, t_img *img, int color)
+{
+	size_t	offset;
+
+	offset = (y * img->line_length) + x * (img->bpp / 8);
+	*(unsigned int *)(img->addr + offset) = color;
+}
+
+void render_pixel(t_img *img, int x, int y, t_object *object)
+{
+	if (object)
+		ft_pixel_put(x, y, img, convert_color(object->color));
+	else
+		ft_pixel_put(x, y, img, 0x000000); 
+}
+
+int	convert_color(t_color color)
+{
+	int r = (int)color.r;
+	int g = (int)color.g;
+	int b = (int)color.b;
+
+	return ((r << 16) | (g << 8) | b);
+}
