+++ test/parse/rt_file/axis_cylinder.rt
@@ -1,0 +1,6 @@
+A 0.2 255,255,255
+C 0,50,100 0,0,-1 80
+L 200,200,200 0.8 255,255,255
+cy 0,0,0 1,0,0 0.02 100 255,0,0
+cy 0,0,0 0,1,0 0.02 100 0,255,0
+cy 0,0,0 0,0,1 0.02 100 0,0,255
