+++ test/parse/rt_file/sphere_and_plane.rt
@@ -1,0 +1,5 @@
+A 0.2 255,255,255
+C 0,0,50 0,0,-1 60
+L 10,10,30 0.7 255,255,255
+pl 0,0,0 0,0,1 200,200,200
+sp 0,0,20 5 255,0,0
